requirements.txt
langchain==0.3.25
langchain-openai==0.3.24
pandas==2.3.0
numpy==2.2.6
python-dotenv==1.1.0
pydantic==2.11.7